back/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
graphene-django>=3.1.5
django-cors-headers>=4.3
Pillow>=10.0
celery>=5.3
redis>=5.0
